# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Përgjithmonë Photobook Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/generate-pdf:
    post:
      operationId: generateProjectPdf
      tags: [projects]
      summary: Generate print-ready PDF for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PdfGenerationResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order from a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/whatsapp:
    get:
      operationId: getOrderWhatsapp
      tags: [orders]
      summary: Get WhatsApp ordering link for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: WhatsApp link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappLinkResponse"

  # ─── UPLOADS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PdfGenerationResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, generating, ready, error]
        pdfUrl:
          type: ["string", "null"]
        shareToken:
          type: ["string", "null"]
        error:
          type: ["string", "null"]

    # ── Orders ────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, userId, projectId, status, priceLek, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        projectId:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, printing, shipped, delivered, cancelled]
        priceLek:
          type: integer
        notes:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [projectId]
      properties:
        projectId:
          type: integer
        notes:
          type: string
    WhatsappLinkResponse:
      type: object
      required: [url, message]
      properties:
        url:
          type: string
        message:
          type: string

    # ── Uploads ───────────────────────────────────────────────────────────
...